Repackage classes for increased readability and ease of navigation
