+++ src/main/java/ginger/exception/IllegalGingerArgumentException.java
@@ -1,3 +1,5 @@
+package ginger.exception;
+
 public class IllegalGingerArgumentException extends Exception {
     public IllegalGingerArgumentException(String message) {
         super(message);
