+++ src/main/java/ginger/exception/IllegalGingerCommandException.java
@@ -1,3 +1,5 @@
+package ginger.exception;
+
 public class IllegalGingerCommandException extends Exception {
     public IllegalGingerCommandException(String message) {
         super(message);
