+++ src/main/java/ginger/task/Deadline.java
@@ -1,3 +1,5 @@
+package ginger.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
