+++ src/main/java/ginger/task/TaskStorage.java
@@ -1,3 +1,5 @@
+package ginger.task;
+
 import java.io.*;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
