+++ src/main/java/ginger/task/ToDo.java
@@ -1,3 +1,5 @@
+package ginger.task;
+
 public class ToDo extends Task {
     public ToDo(String title) {
         super(title);
